Refactor presetRequesthandler and presetQueuehandler
