+++ backend/src/main/java/com/benine/backend/http/RequestHandler.java
@@ -2,9 +2,6 @@ package com.benine.backend.http;
 
 import com.benine.backend.LogEvent;
 import com.benine.backend.Logger;
-import com.benine.backend.camera.CameraController;
-import com.benine.backend.performance.PresetQueueController;
-import com.benine.backend.preset.PresetController;
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.handler.AbstractHandler;
 
@@ -17,12 +14,6 @@ import javax.servlet.http.HttpServletResponse;
  */
 public abstract class RequestHandler extends AbstractHandler {
   
-  private CameraController cameraController;
-  
-  private PresetController presetController;
-  
-  private PresetQueueController presetQueueController;
-  
   private Logger logger;
   
   /**
@@ -30,9 +21,6 @@ public abstract class RequestHandler extends AbstractHandler {
    * @param httpserver to interact with the rest of the system.
    */
   public RequestHandler(HTTPServer httpserver) {
-    this.cameraController = httpserver.getCameraController();
-    this.presetController = httpserver.getPresetController();
-    this.presetQueueController = httpserver.getPresetQueueController();
     this.logger = httpserver.getLogger();
   }
 
@@ -79,29 +67,5 @@ public abstract class RequestHandler extends AbstractHandler {
    */
   protected Logger getLogger() {
     return logger;
-  }
-
-  /**
-   * Returns the cameraController
-   * @return A camera controller object.
-   */
-  protected CameraController getCameraController() {
-    return cameraController;
-  }
-
-  /**
-   * Returns the presetController.
-   * @return presetController to change the presets
-   */
-  protected PresetController getPresetController() {
-    return presetController;
-  }
-
-  /**
-   * Returns the preset queue controller.
-   * @return Preset queues controller
-   */
-  protected PresetQueueController getPresetQueueController() {
-    return presetQueueController;
   }
 }
