+++ backend/src/main/java/com/benine/backend/http/presethandlers/PresetsHandler.java
@@ -14,8 +14,6 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-
-
 public class PresetsHandler extends PresetRequestHandler {
 
   /**
