+++ backend/src/main/java/com/benine/backend/http/presetqueue/PresetQueueRequestHandler.java
@@ -4,6 +4,8 @@ import com.benine.backend.Logger;
 import com.benine.backend.http.HTTPServer;
 import com.benine.backend.http.RequestHandler;
 import com.benine.backend.performance.PresetQueueController;
+import com.benine.backend.preset.Preset;
+import com.benine.backend.preset.PresetController;
 
 import org.eclipse.jetty.server.Request;
 
@@ -14,6 +16,12 @@ import java.util.regex.Pattern;
  * Handles all request requiring the presetsQueue ID.
  */
 public abstract class PresetQueueRequestHandler extends RequestHandler {
+  
+  private Logger logger;
+  
+  private PresetQueueController presetQueueController;
+  
+  private PresetController presetController;
 
   /**
    * PresetQueueRequest handler for the httpserver.
@@ -21,6 +29,9 @@ public abstract class PresetQueueRequestHandler extends RequestHandler {
    */
   public PresetQueueRequestHandler(HTTPServer httpserver) {
     super(httpserver);
+    logger = httpserver.getLogger();
+    presetQueueController = httpserver.getPresetQueueController();
+    presetController = httpserver.getPresetController();
   }
 
   /**
@@ -48,14 +59,21 @@ public abstract class PresetQueueRequestHandler extends RequestHandler {
     return path.replaceFirst(".*/(\\d*)/", "");
   }
   
-  @Override
-  protected Logger getLogger() {
-    return super.getLogger();
+  /**
+   * Get the preset with preset id.
+   * @param presetID to find the preset.
+   * @return the right preset.
+   */
+  protected Preset getPresetById(int presetID) {
+    return presetController.getPresetById(presetID);
   }
   
-  @Override
+  protected Logger getLogger() {
+    return logger;
+  }
+  
   protected PresetQueueController getPresetQueueController() {
-    return super.getPresetQueueController();
+    return presetQueueController;
   }
 
 }
